Link for the Shape Detection API
After listening to this segment, I noticed you don't have a link for it, and also noticed that there's more to read on the API going off of your mention of it finding "landmarks".  To that end, I added the link for the Face Detector section in the WICG spec for the Shape Detection API.

P.S.: loving those initial links for each timestamp, and their per-second precision! Much easier to see what each time block is more or less about at a glance, as well as precisely when they start.